dma: overflow prevention
To get total bytes transfer dma driver multiplies
total number of interrupts and period_size.
Both these variables are defined as uint32.

Currently result of multiplication is typecasted
to uint64, however correct approach should be
to first typecast the variables to uint64 followed
by multiplication. Change corrects this logic so
that for long duration use cases correct position
is returned to ALSA.

Bug 2825875
Bug 200463529

Reviewed-on: https://git-master.nvidia.com/r/c/linux-4.9/+/2288090
(cherry picked from commit 946578db06d4ebedefe4b8187ac97d3edc027be3)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-4.9/+/2289085
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit

+++ drivers/dma/tegra210-adma.c
@@ -1,7 +1,7 @@
 /*
  * ADMA driver for Nvidia's Tegra210 ADMA controller.
  *
- * Copyright (c) 2016-2019, NVIDIA CORPORATION.  All rights reserved.
+ * Copyright (c) 2016-2020, NVIDIA CORPORATION.  All rights reserved.
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms and conditions of the GNU General Public License,
@@ -582,7 +582,8 @@ static unsigned int tegra_adma_get_residue(struct tegra_adma_chan *tdc)
 	/* get transferred data count */
 	tc_transferred = ch_regs->tc - tc_remain;
 
-	tot_xfer = (uint64_t)(tdc->tx_buf_count * ch_regs->tc) + tc_transferred;
+	tot_xfer = (uint64_t)((uint64_t)tdc->tx_buf_count *
+			(uint64_t)ch_regs->tc) + tc_transferred;
 	tot_xfer %= desc->buf_len;
 
 	return desc->buf_len - tot_xfer;
